Worked on the Dataset_generator

+++ scripts/Dataset_generator.py
@@ -9,14 +9,25 @@ class dataset_generator:
     """
 
     def __init__(self):
+        self.dataset = []
+        self.datapath = "./resouces/"
         pass
 
     def showImage(self, img):
         img = cv2.imread(img, 1)
+        cv2.putText(img, "Do you find this person attractive? (Y/N)", (0, 50), 4, (0, 0, 0))
         cv2.imshow("Image", img)
-        cv2.waitKey(0)
-        cv2.destroyAllWindows()
-
+        responded = False
+        while ~responded:
+            key = cv2.waitKey(0)
+            if key == 'Y':
+                # TODO store positive value
+                cv2.destroyAllWindows()
+                responded = True
+            if key == 'N':
+                # TODO store negative value
+                cv2.destroyAllWindows()
+                responded = True
 
 if __name__ == "__main__":
     dg = dataset_generator
